--- anta/tools/__init__.py
@@ -1,7 +1,0 @@
-# Copyright (c) 2023-2024 Arista Networks, Inc.
-# Use of this source code is governed by the Apache License 2.0
-# that can be found in the LICENSE file.
-"""anta.tools submodule.
-
-Utilities for ANTA codebase.
-"""
